Upwind flux for uncoupled advection added.
Bugfix:
The solver was giving oscillations when being run for cx != 1 and cy != 1.
Fix:
1. The problem was that we were using F_x and F_y to calculate the flux
in the xi-eta domain, we should instead use F_xi = dx_dxi * c_x u and
F_eta = dy_deta * c_y * u.

2. In the surface term, the flux of flux was being calculated.

On running the simulation the wave speed is not as expected. Maybe the
F_xi = F_x and F_eta = F_y. Need to check this.